+++ RoboCat/Src/Main.cpp
@@ -3,6 +3,9 @@
 #include "allegro5.h"
 #include "allegro_primitives.h"
 #include "allegro_color.h"
+#include "RectangleObject.h"
+#include "RainParticle.h"
+#include "CircleClass.h"
 
 #include <thread>
 #include <iostream>
@@ -17,210 +20,11 @@ int FRAMERATE = 16;
 
 #if _WIN32
 
-class RainParticle
-{
-public:
-	RainParticle(float xStart, float yStart, float newRadius, float newR, float newG, float newB, float newA);
-	RainParticle();
-	void Read(InputMemoryBitStream& iStream);
-	void Write(OutputMemoryBitStream& oStream) const;
-	void Draw();
-	void UpdatePos(float xChange, float yChange);
 
-	vector<float> color;
-	vector<float> position;
-	float radius;
-};
 
-RainParticle::RainParticle(float xStart, float yStart, float newRadius, float newR, float newG, float newB, float newA)
-{
-	color.push_back(newR);
-	color.push_back(newG);
-	color.push_back(newB);
-	color.push_back(newA);
 
-	position.push_back(xStart);
-	position.push_back(yStart);
 
-	radius = newRadius;
-}
 
-RainParticle::RainParticle()
-{
-	float tempR, tempG, tempB, tempA, tempX, tempY;
-
-	color.push_back(tempR);
-	color.push_back(tempG);
-	color.push_back(tempB);
-	color.push_back(tempA);
-
-	position.push_back(tempX);
-	position.push_back(tempY);
-
-	radius = float();
-}
-
-void RainParticle::Read(InputMemoryBitStream& iStream)
-{
-	iStream.Read(color[0]);
-	iStream.Read(color[1]);
-	iStream.Read(color[2]);
-	iStream.Read(color[3]);
-
-	iStream.Read(position[0]);
-	iStream.Read(position[1]);
-
-	iStream.Read(radius);
-}
-
-void RainParticle::Write(OutputMemoryBitStream& oStream) const
-{
-	oStream.Write(color[0]);
-	oStream.Write(color[1]);
-	oStream.Write(color[2]);
-	oStream.Write(color[3]);
-
-	oStream.Write(position[0]);
-	oStream.Write(position[1]);
-
-	oStream.Write(radius);
-}
-
-void RainParticle::Draw()
-{
-	al_draw_filled_circle(position[0], position[1], radius, al_map_rgba(color[0], color[1], color[2], color[3]));
-}
-
-void RainParticle::UpdatePos(float xChange, float yChange)
-{
-	position[0] += xChange;
-	position[1] += yChange;
-	if (position[0] > windowWidth) position[0] = 0;
-	if (position[1] > windowHeight) position[1] = 0;
-	if (position[0] < 0) position[0] = windowWidth;
-	if (position[1] < 0) position[1] = windowHeight;
-}
-
-class CircleClass
-{
-public:
-	CircleClass(int xStart, int yStart, float newRadius);
-	CircleClass();
-	void Read(InputMemoryBitStream& iStream);
-	void Write(OutputMemoryBitStream& oStream) const;
-	void Draw();
-	void UpdatePos(int xChange, int yChange);
-
-	vector<int> position;
-	float radius;
-};
-
-void CircleClass::Read(InputMemoryBitStream& iStream)
-{
-	iStream.Read(position[0]);
-	iStream.Read(position[1]);
-	iStream.Read(radius);
-}
-
-void CircleClass::Write(OutputMemoryBitStream& oStream) const
-{
-	oStream.Write(position[0]);
-	oStream.Write(position[1]);
-	oStream.Write(radius);
-}
-
-void CircleClass::Draw()
-{
-	al_draw_filled_circle(position[0], position[1], radius, al_map_rgb(0, 255, 0));
-}
-
-CircleClass::CircleClass(int xStart, int yStart, float newRadius)
-{
-	position.push_back(xStart);
-	position.push_back(yStart);
-	radius = newRadius;
-}
-
-CircleClass::CircleClass()
-{
-	int xTemp = 0;
-	int yTemp = 0;
-	position.push_back(xTemp);
-	position.push_back(yTemp);
-	radius = 10;
-}
-
-void CircleClass::UpdatePos(int xChange, int yChange)
-{
-	position[0] += xChange;
-	position[1] += yChange;
-	if (position[0] > windowWidth) position[0] = 0;
-	if (position[1] > windowHeight) position[1] = 0;
-	if (position[0] < 0) position[0] = windowWidth;
-	if (position[1] < 0) position[1] = windowHeight;
-}
-
-class RectangleObject
-{
-public:
-	int width;
-	int height;
-	int xPos;
-	int yPos;
-	void Read(InputMemoryBitStream& stream);
-	void Write(OutputMemoryBitStream& stream) const;
-	RectangleObject(int newWidth, int newHeight, int startPosX, int startPosY);
-	RectangleObject();
-	void UpdatePos(int xChange, int yChange);
-	void Draw();
-	
-};
-
-void RectangleObject::Read(InputMemoryBitStream& stream)
-{
-	stream.Read(width);
-	stream.Read(height);
-	stream.Read(xPos);
-	stream.Read(yPos);
-}
-
-void RectangleObject::Write(OutputMemoryBitStream& stream) const
-{
-	stream.Write(width);
-	stream.Write(height);
-	stream.Write(xPos);
-	stream.Write(yPos);
-}
-
-RectangleObject::RectangleObject(int newWidth, int newHeight, int startPosX, int startPosY)
-{
-	width = newWidth;
-	height = newHeight;
-	xPos = startPosX;
-	yPos = startPosY;
-}
-RectangleObject::RectangleObject()
-{
-	width = 10;
-	height = 10;
-	xPos = rand() % 100 + 1;
-	yPos = rand() % 100 + 1;
-}
-
-void RectangleObject::UpdatePos(int xChange, int yChange)
-{
-	xPos += xChange;
-	yPos += yChange;
-	if (xPos > windowWidth) xPos = 0;
-	if (yPos > windowHeight) yPos = 0;
-	if (xPos < 0) xPos = windowWidth;
-	if (yPos < 0) yPos = windowHeight;
-}
-
-void RectangleObject::Draw()
-{
-	al_draw_filled_rectangle(xPos, yPos, xPos + width, yPos + height, al_map_rgb(255, 0, 0));
-}
 
 void ThrowSocketError(std::string errorMsg)
 {
@@ -238,20 +42,6 @@ ALLEGRO_TIMER* GiveMeATimer(float incr)
 	al_start_timer(timer);
 	return timer;
 }
-
-//class ClientInPacket
-//{
-//public:
-//	ClientPacket();
-//	//~ClientPacket();
-//	CircleClass inCircle;
-//	RainParticle inRain[10];
-//};
-//
-//ClientPacket::ClientPacket()
-//{
-//
-//}
 
 
 TCPSocketPtr StartServer()
@@ -285,6 +75,30 @@ TCPSocketPtr StartServer()
 	return incomingSocket;
 }
 
+TCPSocketPtr StartClientConnection()
+{
+	TCPSocketPtr clientSocket = SocketUtil::CreateTCPSocket(SocketAddressFamily::INET);
+	if (clientSocket == nullptr)
+		ThrowSocketError("Could not create client socket");
+
+	std::string address = StringUtils::Sprintf("127.0.0.1:8081");
+	SocketAddressPtr clientAddress = SocketAddressFactory::CreateIPv4FromString(address.c_str());
+	if (clientAddress == nullptr)
+		ThrowSocketError("Creating client address");
+
+	if (clientSocket->Bind(*clientAddress) != NO_ERROR)
+		ThrowSocketError("Binding Client Socket");
+
+	SocketAddressPtr serverAddress = SocketAddressFactory::CreateIPv4FromString("127.0.0.1:8080");
+	if (serverAddress == nullptr)
+		ThrowSocketError("Creating server address");
+
+	if (clientSocket->Connect(*serverAddress) != NO_ERROR)
+		ThrowSocketError("Connecting to server");
+
+	return clientSocket;
+}
+
 void BradsTotallyOriginalServer()
 {
 	TCPSocketPtr incomingSocket = StartServer();
@@ -292,8 +106,8 @@ void BradsTotallyOriginalServer()
 
 	//Timer
 	ALLEGRO_TIMER* timer = GiveMeATimer(0.001);
-	int currentTime = al_get_timer_count(timer);
-	int lastTime = 0;
+	double currentTime = al_get_timer_count(timer);
+	float lastTime = 0;
 	float deltaTime = 0;
 
 	// Game Loop Variables
@@ -304,17 +118,20 @@ void BradsTotallyOriginalServer()
 	CircleClass greenCircle[numObjects];
 	for (int j = 0; j < numObjects; j++)
 	{
-		greenCircle[j] = CircleClass(100, 100, 20);
+		greenCircle[j] = CircleClass(windowWidth, windowHeight, 100, 100, 20);
 	}
 
 	const int numDroplets = 10;
 	RainParticle rain[numDroplets];
 	for (int k = 0; k < numDroplets; k++)
 	{
-		rain[k] = RainParticle(rand() % windowWidth + 1, rand() % windowHeight + 1, rand() % 5 + 5, 0, 0, rand() % 254 + 1, rand() % 254 + 1);
+		rain[k] = RainParticle(windowWidth, windowHeight, rand() % windowWidth + 1, rand() % windowHeight + 1, rand() % 5 + 5, 0, 0, rand() % 254 + 1, rand() % 254 + 1);
 	}
 
 	RectangleObject inObjects[numObjects];
+	list < RectangleObject > clientProjectiles;
+	list < RectangleObject > serverProjectiles;
+
 
 	ALLEGRO_DISPLAY* display;
 	display = al_create_display(windowWidth, windowHeight);
@@ -441,29 +258,7 @@ void BradsTotallyOriginalServer()
 	al_destroy_display(display);	
 }
 
-TCPSocketPtr StartClientConnection()
-{
-	TCPSocketPtr clientSocket = SocketUtil::CreateTCPSocket(SocketAddressFamily::INET);
-	if (clientSocket == nullptr)
-		ThrowSocketError("Could not create client socket");
 
-	std::string address = StringUtils::Sprintf("127.0.0.1:8081");
-	SocketAddressPtr clientAddress = SocketAddressFactory::CreateIPv4FromString(address.c_str());
-	if (clientAddress == nullptr)
-		ThrowSocketError("Creating client address");
-
-	if (clientSocket->Bind(*clientAddress) != NO_ERROR)
-		ThrowSocketError("Binding Client Socket");
-
-	SocketAddressPtr serverAddress = SocketAddressFactory::CreateIPv4FromString("127.0.0.1:8080");
-	if (serverAddress == nullptr)
-		ThrowSocketError("Creating server address");
-
-	if (clientSocket->Connect(*serverAddress) != NO_ERROR)
-		ThrowSocketError("Connecting to server");
-
-	return clientSocket;
-}
 void BradsLessOriginalClient()
 {
 	TCPSocketPtr clientSocket = StartClientConnection();
@@ -471,8 +266,8 @@ void BradsLessOriginalClient()
 
 	//timer	
 	ALLEGRO_TIMER* timer = GiveMeATimer(0.001);
-	int currentTime = al_get_timer_count(timer);
-	int lastTime = 0;
+	double currentTime = al_get_timer_count(timer);
+	float lastTime = 0;
 	float deltaTime = 0;
 
 	bool exit = false;
@@ -480,18 +275,22 @@ void BradsLessOriginalClient()
 
 	//create Game Objects
 	RectangleObject outRectangles[numObjects];
+	list < RectangleObject > projectiles;
 
 	for (int j = 0; j < numObjects; j++)
 	{
-		outRectangles[j] = RectangleObject(20, 20, rand() % windowWidth + 1, rand() % windowHeight + 1);
+		outRectangles[j] = RectangleObject("clientPlayer", windowWidth, windowHeight, 20, 20, rand() % windowWidth + 1, rand() % windowHeight + 1);
 	}
 
 	// Read In Objects
 	CircleClass inObjects[numObjects];
 
+	list < RectangleObject > clientProjectiles;
+	list < RectangleObject > serverProjectiles;
+
 	const int numDroplets = 10;
 	RainParticle rain[numDroplets];
-
+	
 	// Main Game Loop
 	ALLEGRO_DISPLAY* display;
 	display = al_create_display(windowWidth, windowHeight);
@@ -537,8 +336,7 @@ void BradsLessOriginalClient()
 			for (int j = 0; j < numDroplets; j++)
 			{
 				rain[j].Read(iStream);
-			}
-			
+			}		
 		}		
 
 		ALLEGRO_EVENT events;
@@ -600,6 +398,7 @@ void BradsLessOriginalClient()
 			outRectangles[i].Draw();
 			inObjects[i].Draw();
 		}
+
 		for (int j = 0; j < numDroplets; j++)
 		{
 			rain[j].Draw();
